requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.0
PyJWT==2.9.0
redis==5.2.0
rq==1.16.2
transformers==4.44.2
email-validator==2.2.0
python-multipart==0.0.9
alembic==1.13.2
psycopg[binary]>=3.2